got all expenses from backend and save it in room database

+++ app/src/main/java/com/expensemoitor/expensemonitor/createnewexpense/CreateNewExpenseFragmentViewModel.kt
@@ -73,7 +73,7 @@ class CreateNewExpenseFragmentViewModel(var application: Application) : ViewMode
     private fun createNewExpense(amount:String,description:String,date:String,category:String){
         viewModelScope.launch {
             val expenseData = getCurrencyFromSettings()?.let {
-                ExpenseData(amount.toBigDecimal(),description,date,
+                ExpenseData(amount,description,date,
                     it,category)
             }
             val getResponse = expenseData?.let {
